+++ source/select.js
@@ -784,8 +784,8 @@ export default class Select extends PureComponent
 					selected_style_classes,
 					'rrui__select__button',
 					{
-						'rrui__select__button--empty' : !selected_label,
-						'rrui__input-field--invalid'  : this.should_indicate_invalid()
+						'rrui__select__button--empty'   : !selected_label,
+						'rrui__select__button--invalid' : this.should_indicate_invalid()
 					}
 				) }>