+++ source/select.js
@@ -736,6 +736,7 @@ export default class Select extends PureComponent
 
 		const selected_style_classes =
 		{
+			'rrui__input-element',
 			// CSS selector performance optimization
 			'rrui__select__selected--disabled' : disabled
 		}
@@ -785,7 +786,8 @@ export default class Select extends PureComponent
 					'rrui__select__button',
 					{
 						'rrui__select__button--empty'   : !selected_label,
-						'rrui__select__button--invalid' : this.should_indicate_invalid()
+						'rrui__select__button--invalid' : this.should_indicate_invalid(),
+						'rrui__input-element--invalid'  : this.should_indicate_invalid()
 					}
 				) }>